Restructures outline
One issue I had today was the install process taking various amounts of
time and users being split between GUI and CLI. To address this I'm
changing the workshop to be more structured lecture with more up-front
examples before switching to hands-on activities.

Another issue was users being unsure of what to type, so in the
hands-on section I'll provide more guidance to help them know exactly
what to enter.